+++ src/components/windows/magic_f_gui.py
@@ -1,64 +1,74 @@
 import tkinter as tk
 from tkinter import ttk
+from src.components.frames.title_frame import TitleFrame
 from typing import TYPE_CHECKING
 
-
-from src.components.frames.base_frame import BaseFrame
-from src.components.frames.title_frame import TitleFrame
-
+from src.components.frames.configurable_frame import ConfigurableFrame
 if TYPE_CHECKING:
     from src.controllers.magic_f import MagicF
 
-
-
-class MagicFGUI(BaseFrame):
+class MagicFGUI(ConfigurableFrame):
     
-    def __init__(self, magic_f: 'MagicF', master, app_controller, loop):
-        super().__init__(master)
+    def __init__(
+        self,
+        magic_f: 'MagicF',
+        master,
+        app_controller,
+    ):
+        super().__init__(master=master)
+        
+        self.master = master
         self.app_controller = app_controller
         self.service_controller = magic_f
-        self.config = self.service_controller.app_config
         
-        self.loop = loop
-        self.subservices = self.service_controller.subservices
-        self.current_subservice_screen = None
-        self.original_pack_configs = {}
+        self.config = self.service_controller.config
+        self.entries = {}
+        self.toggle_key_label = None
+        self.selected_option = None
         self.init_gui()
-
         
     
     def init_gui(self):
+        
         title_frame = TitleFrame(
             self,
             self.service_controller.__name__()
         )
         title_frame.pack(side=tk.TOP, fill=tk.X)
         
+        self.toggle_key_label = tk.Label(self, text=f"Select mode to enable when pressing {(self.config.toggle_key).upper()}", font=("Arial", 8, "italic"), foreground="#800000")
+        self.toggle_key_label.pack(pady=20)
+        
+        self.selection_frame = ttk.Frame(self)
+        self.selection_frame.place(relx=0.5, rely=0.5, anchor='center')
 
-        self.create_buttons()
-        self.store_widget_configs()
+        self.selected_option = tk.StringVar(value="veggies")
+        for name, value in self.service_controller.options.items():
+            if name == "Dumper" or name == "Crafter":
+                frame = ttk.Frame(self.selection_frame)
+                frame.pack(anchor='w')
 
+                radiobutton = ttk.Radiobutton(
+                    frame,
+                    text=name,
+                    variable=self.selected_option,
+                    value=value,
+                )
+                radiobutton.pack(side=tk.LEFT, pady=2)
+
+                self.entries[name] = ttk.Entry(frame, width=10)
+                self.entries[name].pack(side=tk.LEFT, padx=10)
+            else:
+                radiobutton = ttk.Radiobutton(
+                    self.selection_frame,
+                    text=name,
+                    variable=self.selected_option,
+                    value=value
+                )
+                radiobutton.pack(anchor='w')
+            
     
     def destroy_gui(self):
+        self.service_controller.destroy()
         super().destroy()
         self.app_controller.show_main()
-    
-    
-    def store_widget_configs(self):
-        """
-        Stores the pack configurations of all widgets.
-        """
-        for widget in self.winfo_children():
-            self.original_pack_configs[widget] = widget.pack_info()
-
-    def create_buttons(self):
-        for key in self.subservices:
-            button = ttk.Button(self, text=key, command=self.show_subservice_command(app_name=key))
-            button.pack(side=tk.TOP, pady=10)
-        
-    def show_subservice_command(self, app_name):
-        def command():
-            self.service_controller.show_subservice(app_name)
-        return command
-    
-