+++ CHANGELOG.md
@@ -81,3 +81,19 @@
 ## [1.3.1] - 03/06/2024
 
 - Changed label colors and descriptions
+
+## [1.4.0] - 04/06/2024
+
+- Implement Magic-F - Veggies Mode
+
+## [1.5.0] - 04/06/2024
+
+- Implement Magic-F - Retrieve
+
+## [1.6.0] - 04/06/2024
+
+- Implement Magic-F - Dump
+
+## [1.7.0] - 04/06/2024
+
+- Implement Magic-F - Craft